Move WholeLinesIter to whole_lines.rs
